fix solution
