+++ src/Root.tsx
@@ -1,7 +1,11 @@
-import { HashRouter as Router, Routes, Route } from 'react-router-dom';
+import {
+  HashRouter as Router,
+  Routes,
+  Route,
+  Navigate,
+} from 'react-router-dom';
 import { App } from './App';
 import { HomePage } from './components/HomePage';
-import { Tab } from './components/Tab';
 import { TabsPage } from './components/TabsPage';
 import { PageNotFound } from './components/PageNotFound';
 
@@ -11,8 +15,9 @@ export const Root = () => (
       <Route path="/" element={<App />}>
         <Route index element={<HomePage />}></Route>
         <Route path="tabs" element={<TabsPage />}>
-          <Route path=":tabId?" element={<Tab />}></Route>
+          <Route path=":tabId?" element={<TabsPage />}></Route>
         </Route>
+        <Route path="home" element={<Navigate to="/" replace/>}></Route>
         <Route path="*" element={<PageNotFound />}></Route>
       </Route>
     </Routes>
