+++ src/api/tabs-api.ts
@@ -1,4 +1,6 @@
-export const tabs = [
+import { TabT } from "../types/Tab";
+
+export const tabs: TabT[] = [
   { id: 'tab-1', title: 'Tab 1', content: 'Some text 1' },
   { id: 'tab-2', title: 'Tab 2', content: 'Some text 2' },
   { id: 'tab-3', title: 'Tab 3', content: 'Some text 3' },
