--- src/components/Tab.tsx
@@ -1,22 +1,0 @@
-import { useParams } from 'react-router-dom';
-import { tabs } from '../api/tabs-api';
-
-export const Tab = () => {
-  const { tabId } = useParams();
-
-  const activeTab = tabs.find(tab => tab.id === tabId);
-
-  return (
-    <>
-      {activeTab ? (
-        <div className="block" data-cy="TabContent">
-          {activeTab.content}
-        </div>
-      ) : (
-        <div className="block" data-cy="TabContent">
-          Please select a tab
-        </div>
-      )}
-    </>
-  );
-};
