+++ src/components/TabsPage.tsx
@@ -1,13 +1,10 @@
-import { Link, Outlet, useParams } from 'react-router-dom';
-
-const tabs = [
-  { id: 'tab-1', title: 'Tab 1', content: 'Some text 1' },
-  { id: 'tab-2', title: 'Tab 2', content: 'Some text 2' },
-  { id: 'tab-3', title: 'Tab 3', content: 'Some text 3' },
-];
+import { Link, useParams } from 'react-router-dom';
+import { tabs } from '../api/tabs-api';
 
 export const TabsPage = () => {
   const { tabId } = useParams();
+
+  const activeTab = tabs.find(tab => tab.id === tabId);
 
   return (
     <>
@@ -20,7 +17,7 @@ export const TabsPage = () => {
               data-cy="Tab"
               className={tab.id === tabId ? 'is-active' : ''}
             >
-              <Link to={`${tab.id}`} data-cy="TabLink">
+              <Link to={tab.id} data-cy="TabLink">
                 {tab.title}
               </Link>
             </li>
@@ -28,7 +25,15 @@ export const TabsPage = () => {
         </ul>
       </div>
 
-      <Outlet />
+      {activeTab ? (
+        <div className="block" data-cy="TabContent">
+          {activeTab.content}
+        </div>
+      ) : (
+        <div className="block" data-cy="TabContent">
+          Please select a tab
+        </div>
+      )}
     </>
   );
 };
